refactor(navigator): NavigatorMain to Stack
